keep from polars.utils.udfs import _get_shared_lib_location  working but issue deprecationwarning
