+++ py-polars/polars/utils/udfs.py
@@ -1,15 +1,14 @@
-import os
 from typing import Any
+
+from polars._utils.deprecation import issue_deprecation_warning
+from polars.api import get_shared_lib_location
 
 __all__ = ["_get_shared_lib_location"]
 
 
 def _get_shared_lib_location(main_file: Any) -> str:
-    directory = os.path.dirname(main_file)  # noqa: PTH120
-    return os.path.join(  # noqa: PTH118
-        directory, next(filter(_is_shared_lib, os.listdir(directory)))
+    issue_deprecation_warning(
+        "`polars.utils` is deprecated and will be made private. Use `from polars.api import get_shared_lib_location` instead.",
+        version="0.20.14",
     )
-
-
-def _is_shared_lib(file: str) -> bool:
-    return file.endswith((".so", ".dll", ".pyd"))
+    return get_shared_lib_location(main_file)
